+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added myems-admin.conf and myems-web.conf for nginx
 - added binding microgrid to space in myems-admin
 - added blank page image for space mnvironment monitor in myems-web
+- added quick mode for svg in myems-admin and myems-api
 ### Changed
 - subspace names under the same parent space cannot be duplicated
 - updated power stations